+++ src/downloader2.rs
@@ -43,7 +43,13 @@ use tokio_util::task::AbortOnDropHandle;
 use tracing::{debug, error, info, trace};
 
 mod planners;
-pub(crate) use planners::*;
+use planners::*;
+
+mod state;
+use state::*;
+
+mod actor;
+use actor::*;
 
 #[derive(
     Debug, Clone, Copy, Serialize, Deserialize, PartialEq, Eq, PartialOrd, Ord, derive_more::From,
@@ -146,35 +152,6 @@ pub enum ObserveEvent {
     },
 }
 
-/// Global information about a peer
-#[derive(Debug, Default)]
-struct PeerState {
-    /// Executed downloads, to calculate the average download speed.
-    ///
-    /// This gets updated as soon as possible when new data has been downloaded.
-    download_history: VecDeque<(Duration, (u64, u64))>,
-}
-
-/// Information about one blob on one peer
-struct PeerBlobState {
-    /// The subscription id for the subscription
-    subscription_id: BitfieldSubscriptionId,
-    /// The number of subscriptions this peer has
-    subscription_count: usize,
-    /// chunk ranges this peer reports to have
-    ranges: ChunkRanges,
-}
-
-impl PeerBlobState {
-    fn new(subscription_id: BitfieldSubscriptionId) -> Self {
-        Self {
-            subscription_id,
-            subscription_count: 1,
-            ranges: ChunkRanges::empty(),
-        }
-    }
-}
-
 /// A download request
 #[derive(Debug, Clone)]
 pub struct DownloadRequest {
@@ -193,25 +170,6 @@ pub struct ObserveRequest {
     pub ranges: ChunkRanges,
     /// Buffer size
     pub buffer: usize,
-}
-
-struct DownloadState {
-    /// The request this state is for
-    request: DownloadRequest,
-    /// Ongoing downloads
-    peer_downloads: BTreeMap<NodeId, PeerDownloadState>,
-    /// Set to true if the download needs rebalancing
-    needs_rebalancing: bool,
-}
-
-impl DownloadState {
-    fn new(request: DownloadRequest) -> Self {
-        Self {
-            request,
-            peer_downloads: BTreeMap::new(),
-            needs_rebalancing: false,
-        }
-    }
 }
 
 #[derive(Debug)]
@@ -240,171 +198,6 @@ where
     }
 }
 
-/// Wrapper for the downloads map
-///
-/// This is so we can later optimize access by fields other than id, such as hash.
-#[derive(Default)]
-struct Downloads {
-    by_id: BTreeMap<DownloadId, DownloadState>,
-}
-
-impl Downloads {
-    fn remove(&mut self, id: &DownloadId) -> Option<DownloadState> {
-        self.by_id.remove(id)
-    }
-
-    fn contains_key(&self, id: &DownloadId) -> bool {
-        self.by_id.contains_key(id)
-    }
-
-    fn insert(&mut self, id: DownloadId, state: DownloadState) {
-        self.by_id.insert(id, state);
-    }
-
-    fn iter_mut_for_hash(
-        &mut self,
-        hash: Hash,
-    ) -> impl Iterator<Item = (&DownloadId, &mut DownloadState)> {
-        self.by_id
-            .iter_mut()
-            .filter(move |x| x.1.request.hash == hash)
-    }
-
-    fn iter(&mut self) -> impl Iterator<Item = (&DownloadId, &DownloadState)> {
-        self.by_id.iter()
-    }
-
-    /// Iterate over all downloads for a given hash
-    fn values_for_hash(&self, hash: Hash) -> impl Iterator<Item = &DownloadState> {
-        self.by_id.values().filter(move |x| x.request.hash == hash)
-    }
-
-    fn values_mut_for_hash(&mut self, hash: Hash) -> impl Iterator<Item = &mut DownloadState> {
-        self.by_id
-            .values_mut()
-            .filter(move |x| x.request.hash == hash)
-    }
-
-    fn by_id_mut(&mut self, id: DownloadId) -> Option<&mut DownloadState> {
-        self.by_id.get_mut(&id)
-    }
-
-    fn by_peer_download_id_mut(
-        &mut self,
-        id: PeerDownloadId,
-    ) -> Option<(&DownloadId, &mut DownloadState)> {
-        self.by_id
-            .iter_mut()
-            .filter(|(_, v)| v.peer_downloads.iter().any(|(_, state)| state.id == id))
-            .next()
-    }
-}
-
-#[derive(Default)]
-struct Bitfields {
-    // Counters to generate unique ids for various requests.
-    // We could use uuid here, but using integers simplifies testing.
-    //
-    // the next subscription id
-    subscription_id_gen: IdGenerator<BitfieldSubscriptionId>,
-    by_peer_and_hash: BTreeMap<(BitfieldPeer, Hash), PeerBlobState>,
-}
-
-impl Bitfields {
-    fn retain<F>(&mut self, mut f: F)
-    where
-        F: FnMut(&(BitfieldPeer, Hash), &mut PeerBlobState) -> bool,
-    {
-        self.by_peer_and_hash.retain(|k, v| f(k, v));
-    }
-
-    fn get(&self, key: &(BitfieldPeer, Hash)) -> Option<&PeerBlobState> {
-        self.by_peer_and_hash.get(key)
-    }
-
-    fn get_local(&self, hash: Hash) -> Option<&PeerBlobState> {
-        self.by_peer_and_hash.get(&(BitfieldPeer::Local, hash))
-    }
-
-    fn get_mut(&mut self, key: &(BitfieldPeer, Hash)) -> Option<&mut PeerBlobState> {
-        self.by_peer_and_hash.get_mut(key)
-    }
-
-    fn get_local_mut(&mut self, hash: Hash) -> Option<&mut PeerBlobState> {
-        self.by_peer_and_hash.get_mut(&(BitfieldPeer::Local, hash))
-    }
-
-    fn insert(&mut self, key: (BitfieldPeer, Hash), value: PeerBlobState) {
-        self.by_peer_and_hash.insert(key, value);
-    }
-
-    fn contains_key(&self, key: &(BitfieldPeer, Hash)) -> bool {
-        self.by_peer_and_hash.contains_key(key)
-    }
-
-    fn remote_for_hash(&self, hash: Hash) -> impl Iterator<Item = (&NodeId, &PeerBlobState)> {
-        self.by_peer_and_hash
-            .iter()
-            .filter_map(move |((peer, h), state)| {
-                if let BitfieldPeer::Remote(peer) = peer {
-                    if *h == hash {
-                        Some((peer, state))
-                    } else {
-                        None
-                    }
-                } else {
-                    None
-                }
-            })
-    }
-}
-
-struct PeerDownloadState {
-    id: PeerDownloadId,
-    ranges: ChunkRanges,
-}
-
-struct DownloaderState {
-    // all peers I am tracking for any download
-    peers: BTreeMap<NodeId, PeerState>,
-    // all bitfields I am tracking, both for myself and for remote peers
-    //
-    // each item here corresponds to an active subscription
-    bitfields: Bitfields,
-    /// Observers for local bitfields
-    observers: Observers,
-    // all active downloads
-    //
-    // these are user downloads. each user download gets split into one or more
-    // peer downloads.
-    downloads: Downloads,
-    // discovery tasks
-    //
-    // there is a discovery task for each blob we are interested in.
-    discovery: BTreeMap<Hash, DiscoveryId>,
-    // the next discovery id
-    discovery_id_gen: IdGenerator<DiscoveryId>,
-    // the next peer download id
-    peer_download_id_gen: IdGenerator<PeerDownloadId>,
-    // the download planner
-    planner: Box<dyn DownloadPlanner>,
-}
-
-impl DownloaderState {
-    fn new(planner: Box<dyn DownloadPlanner>) -> Self {
-        Self {
-            peers: Default::default(),
-            downloads: Default::default(),
-            bitfields: Default::default(),
-            discovery: Default::default(),
-            observers: Default::default(),
-            discovery_id_gen: Default::default(),
-            peer_download_id_gen: Default::default(),
-            planner,
-        }
-    }
-}
-
 /// Peer for a bitfield subscription
 #[derive(Debug, PartialEq, Eq, PartialOrd, Ord, Clone, Copy)]
 pub enum BitfieldPeer {
@@ -412,648 +205,6 @@ pub enum BitfieldPeer {
     Local,
     /// A bitfield from a remote peer
     Remote(NodeId),
-}
-
-#[derive(Debug)]
-enum Command {
-    /// A user request to start a download.
-    StartDownload {
-        /// The download request
-        request: DownloadRequest,
-        /// The unique id, to be assigned by the caller
-        id: DownloadId,
-    },
-    /// A user request to abort a download.
-    StopDownload { id: DownloadId },
-    /// A full bitfield for a blob and a peer
-    Bitfield {
-        /// The peer that sent the bitfield.
-        peer: BitfieldPeer,
-        /// The blob for which the bitfield is
-        hash: Hash,
-        /// The complete bitfield
-        ranges: ChunkRanges,
-    },
-    /// An update of a bitfield for a hash
-    ///
-    /// This is used both to update the bitfield of remote peers, and to update
-    /// the local bitfield.
-    BitfieldUpdate {
-        /// The peer that sent the update.
-        peer: BitfieldPeer,
-        /// The blob that was updated.
-        hash: Hash,
-        /// The ranges that were added
-        added: ChunkRanges,
-        /// The ranges that were removed
-        removed: ChunkRanges,
-    },
-    /// A chunk was downloaded, but not yet stored
-    ///
-    /// This can only be used for updating peer stats, not for completing downloads.
-    ChunksDownloaded {
-        /// Time when the download was received
-        time: Duration,
-        /// The peer that sent the chunk
-        peer: NodeId,
-        /// The blob that was downloaded
-        hash: Hash,
-        /// The ranges that were added locally
-        added: ChunkRanges,
-    },
-    /// A peer download has completed
-    PeerDownloadComplete {
-        id: PeerDownloadId,
-        #[allow(dead_code)]
-        result: anyhow::Result<Stats>,
-    },
-    /// Stop tracking a peer for all blobs, for whatever reason
-    #[allow(dead_code)]
-    DropPeer { peer: NodeId },
-    /// A peer has been discovered
-    PeerDiscovered { peer: NodeId, hash: Hash },
-    ///
-    ObserveLocal {
-        id: ObserveId,
-        hash: Hash,
-        ranges: ChunkRanges,
-    },
-    ///
-    StopObserveLocal { id: ObserveId },
-    /// A tick from the driver, for rebalancing
-    Tick { time: Duration },
-}
-
-#[derive(Debug, PartialEq, Eq)]
-enum Event {
-    SubscribeBitfield {
-        peer: BitfieldPeer,
-        hash: Hash,
-        /// The unique id of the subscription
-        id: BitfieldSubscriptionId,
-    },
-    UnsubscribeBitfield {
-        /// The unique id of the subscription
-        id: BitfieldSubscriptionId,
-    },
-    LocalBitfield {
-        id: ObserveId,
-        ranges: ChunkRanges,
-    },
-    LocalBitfieldUpdate {
-        id: ObserveId,
-        added: ChunkRanges,
-        removed: ChunkRanges,
-    },
-    StartDiscovery {
-        hash: Hash,
-        /// The unique id of the discovery task
-        id: DiscoveryId,
-    },
-    StopDiscovery {
-        /// The unique id of the discovery task
-        id: DiscoveryId,
-    },
-    StartPeerDownload {
-        /// The unique id of the peer download task
-        id: PeerDownloadId,
-        peer: NodeId,
-        hash: Hash,
-        ranges: ChunkRanges,
-    },
-    StopPeerDownload {
-        /// The unique id of the peer download task
-        id: PeerDownloadId,
-    },
-    DownloadComplete {
-        /// The unique id of the user download
-        id: DownloadId,
-    },
-    /// An error that stops processing the command
-    Error {
-        message: String,
-    },
-}
-
-impl DownloaderState {
-    /// Apply a command and return the events that were generated
-    fn apply(&mut self, cmd: Command) -> Vec<Event> {
-        let mut evs = vec![];
-        self.apply_mut(cmd, &mut evs);
-        evs
-    }
-
-    /// Apply a command, using a mutable reference to the events
-    fn apply_mut(&mut self, cmd: Command, evs: &mut Vec<Event>) {
-        if let Err(cause) = self.apply_mut_0(cmd, evs) {
-            evs.push(Event::Error {
-                message: format!("{cause}"),
-            });
-        }
-    }
-
-    /// Stop a download and clean up
-    ///
-    /// This is called both for stopping a download before completion, and for
-    /// cleaning up after a successful download.
-    ///
-    /// Cleanup involves emitting events for
-    /// - stopping all peer downloads
-    /// - unsubscribing from bitfields if needed
-    /// - stopping the discovery task if needed
-    fn stop_download(&mut self, id: DownloadId, evs: &mut Vec<Event>) -> anyhow::Result<()> {
-        let removed = self
-            .downloads
-            .remove(&id)
-            .context(format!("removed unknown download {id:?}"))?;
-        let removed_hash = removed.request.hash;
-        // stop associated peer downloads
-        for peer_download in removed.peer_downloads.values() {
-            evs.push(Event::StopPeerDownload {
-                id: peer_download.id,
-            });
-        }
-        // unsubscribe from bitfields that have no more subscriptions
-        self.bitfields.retain(|(_peer, hash), state| {
-            if *hash == removed_hash {
-                state.subscription_count -= 1;
-                if state.subscription_count == 0 {
-                    evs.push(Event::UnsubscribeBitfield {
-                        id: state.subscription_id,
-                    });
-                    return false;
-                }
-            }
-            true
-        });
-        let hash_interest = self.downloads.values_for_hash(removed.request.hash).count();
-        if hash_interest == 0 {
-            // stop the discovery task if we were the last one interested in the hash
-            let discovery_id = self
-                .discovery
-                .remove(&removed.request.hash)
-                .context(format!(
-                    "removed unknown discovery task for {}",
-                    removed.request.hash
-                ))?;
-            evs.push(Event::StopDiscovery { id: discovery_id });
-        }
-        Ok(())
-    }
-
-    /// Apply a command and bail out on error
-    fn apply_mut_0(&mut self, cmd: Command, evs: &mut Vec<Event>) -> anyhow::Result<()> {
-        trace!("handle_command {cmd:?}");
-        match cmd {
-            Command::StartDownload { request, id } => {
-                // ids must be uniquely assigned by the caller!
-                anyhow::ensure!(
-                    !self.downloads.contains_key(&id),
-                    "duplicate download request {id:?}"
-                );
-                let hash = request.hash;
-                // either we have a subscription for this blob, or we have to create one
-                if let Some(state) = self.bitfields.get_local_mut(hash) {
-                    // just increment the count
-                    state.subscription_count += 1;
-                } else {
-                    // create a new subscription
-                    let subscription_id = self.bitfields.subscription_id_gen.next();
-                    evs.push(Event::SubscribeBitfield {
-                        peer: BitfieldPeer::Local,
-                        hash,
-                        id: subscription_id,
-                    });
-                    self.bitfields.insert(
-                        (BitfieldPeer::Local, hash),
-                        PeerBlobState::new(subscription_id),
-                    );
-                }
-                if !self.discovery.contains_key(&request.hash) {
-                    // start a discovery task
-                    let id = self.discovery_id_gen.next();
-                    evs.push(Event::StartDiscovery { hash, id });
-                    self.discovery.insert(request.hash, id);
-                }
-                self.downloads.insert(id, DownloadState::new(request));
-                self.check_completion(hash, Some(id), evs)?;
-                self.start_downloads(hash, Some(id), evs)?;
-            }
-            Command::PeerDownloadComplete { id, .. } => {
-                let Some((download_id, download)) = self.downloads.by_peer_download_id_mut(id)
-                else {
-                    // the download was already removed
-                    return Ok(());
-                };
-                let download_id = *download_id;
-                let hash = download.request.hash;
-                download.peer_downloads.retain(|_, v| v.id != id);
-                self.start_downloads(hash, Some(download_id), evs)?;
-            }
-            Command::StopDownload { id } => {
-                self.stop_download(id, evs)?;
-            }
-            Command::PeerDiscovered { peer, hash } => {
-                if self
-                    .bitfields
-                    .contains_key(&(BitfieldPeer::Remote(peer), hash))
-                {
-                    // we already have a subscription for this peer
-                    return Ok(());
-                };
-                // check if anybody needs this peer
-                if self.downloads.values_for_hash(hash).next().is_none() {
-                    return Ok(());
-                }
-                // create a peer state if it does not exist
-                let _state = self.peers.entry(peer).or_default();
-                // create a new subscription
-                let subscription_id = self.bitfields.subscription_id_gen.next();
-                evs.push(Event::SubscribeBitfield {
-                    peer: BitfieldPeer::Remote(peer),
-                    hash,
-                    id: subscription_id,
-                });
-                self.bitfields.insert(
-                    (BitfieldPeer::Remote(peer), hash),
-                    PeerBlobState::new(subscription_id),
-                );
-            }
-            Command::DropPeer { peer } => {
-                self.bitfields.retain(|(p, _), state| {
-                    if *p == BitfieldPeer::Remote(peer) {
-                        // todo: should we emit unsubscribe evs here?
-                        evs.push(Event::UnsubscribeBitfield {
-                            id: state.subscription_id,
-                        });
-                        return false;
-                    } else {
-                        return true;
-                    }
-                });
-                self.peers.remove(&peer);
-            }
-            Command::Bitfield { peer, hash, ranges } => {
-                let state = self.bitfields.get_mut(&(peer, hash)).context(format!(
-                    "bitfields for unknown peer {peer:?} and hash {hash}"
-                ))?;
-                let _chunks = total_chunks(&ranges).context("open range")?;
-                if peer == BitfieldPeer::Local {
-                    // we got a new local bitmap, notify local observers
-                    // we must notify all local observers, even if the bitmap is empty
-                    if let Some(observers) = self.observers.get_by_hash(&hash) {
-                        for (id, request) in observers {
-                            let ranges = &ranges & &request.ranges;
-                            evs.push(Event::LocalBitfield { id: *id, ranges });
-                        }
-                    }
-                    state.ranges = ranges;
-                    self.check_completion(hash, None, evs)?;
-                } else {
-                    // We got an entirely new peer, mark all affected downloads for rebalancing
-                    for download in self.downloads.values_mut_for_hash(hash) {
-                        if ranges.intersects(&download.request.ranges) {
-                            download.needs_rebalancing = true;
-                        }
-                    }
-                    state.ranges = ranges;
-                }
-                // we have to call start_downloads even if the local bitfield set, since we don't know in which order local and remote bitfields arrive
-                self.start_downloads(hash, None, evs)?;
-            }
-            Command::BitfieldUpdate {
-                peer,
-                hash,
-                added,
-                removed,
-            } => {
-                let state = self.bitfields.get_mut(&(peer, hash)).context(format!(
-                    "bitfield update for unknown peer {peer:?} and hash {hash}"
-                ))?;
-                if peer == BitfieldPeer::Local {
-                    // we got a local bitfield update, notify local observers
-                    // for updates we can just notify the observers that have a non-empty intersection with the update
-                    if let Some(observers) = self.observers.get_by_hash(&hash) {
-                        for (id, request) in observers {
-                            let added = &added & &request.ranges;
-                            let removed = &removed & &request.ranges;
-                            if !added.is_empty() || !removed.is_empty() {
-                                evs.push(Event::LocalBitfieldUpdate {
-                                    id: *id,
-                                    added: &added & &request.ranges,
-                                    removed: &removed & &request.ranges,
-                                });
-                            }
-                        }
-                    }
-                    state.ranges |= added;
-                    state.ranges &= !removed;
-                    self.check_completion(hash, None, evs)?;
-                } else {
-                    // We got more data for this hash, mark all affected downloads for rebalancing
-                    for download in self.downloads.values_mut_for_hash(hash) {
-                        // if removed is non-empty, that is so weird that we just rebalance in any case
-                        if !removed.is_empty() || added.intersects(&download.request.ranges) {
-                            download.needs_rebalancing = true;
-                        }
-                    }
-                    state.ranges |= added;
-                    state.ranges &= !removed;
-                    // a local bitfield update does not make more data available, so we don't need to start downloads
-                    self.start_downloads(hash, None, evs)?;
-                }
-            }
-            Command::ChunksDownloaded {
-                time,
-                peer,
-                hash,
-                added,
-            } => {
-                let state = self.bitfields.get_local_mut(hash).context(format!(
-                    "chunks downloaded before having local bitfield for {hash}"
-                ))?;
-                let total_downloaded = total_chunks(&added).context("open range")?;
-                let total_before = total_chunks(&state.ranges).context("open range")?;
-                state.ranges |= added;
-                let total_after = total_chunks(&state.ranges).context("open range")?;
-                let useful_downloaded = total_after - total_before;
-                let peer = self.peers.get_mut(&peer).context(format!(
-                    "performing download before having peer state for {peer}"
-                ))?;
-                peer.download_history
-                    .push_back((time, (total_downloaded, useful_downloaded)));
-            }
-            Command::Tick { time } => {
-                let window = 10;
-                let horizon = time.saturating_sub(Duration::from_secs(window));
-                // clean up download history
-                let mut to_rebalance = vec![];
-                for (peer, state) in self.peers.iter_mut() {
-                    state
-                        .download_history
-                        .retain(|(duration, _)| *duration > horizon);
-                    let mut sum_total = 0;
-                    let mut sum_useful = 0;
-                    for (_, (total, useful)) in state.download_history.iter() {
-                        sum_total += total;
-                        sum_useful += useful;
-                    }
-                    let speed_useful = (sum_useful as f64) / (window as f64);
-                    let speed_total = (sum_total as f64) / (window as f64);
-                    trace!("peer {peer} download speed {speed_total} cps total, {speed_useful} cps useful");
-                }
-
-                for (id, download) in self.downloads.iter() {
-                    if !download.needs_rebalancing {
-                        // nothing has changed that affects this download
-                        continue;
-                    }
-                    let n_peers = self
-                        .bitfields
-                        .remote_for_hash(download.request.hash)
-                        .count();
-                    if download.peer_downloads.len() >= n_peers {
-                        // we are already downloading from all peers for this hash
-                        continue;
-                    }
-                    to_rebalance.push(*id);
-                }
-                for id in to_rebalance {
-                    self.rebalance_download(id, evs)?;
-                }
-            }
-            Command::ObserveLocal { id, hash, ranges } => {
-                // either we have a subscription for this blob, or we have to create one
-                if let Some(state) = self.bitfields.get_local_mut(hash) {
-                    // just increment the count
-                    state.subscription_count += 1;
-                    // emit the current bitfield
-                    evs.push(Event::LocalBitfield {
-                        id,
-                        ranges: state.ranges.clone(),
-                    });
-                } else {
-                    // create a new subscription
-                    let subscription_id = self.bitfields.subscription_id_gen.next();
-                    evs.push(Event::SubscribeBitfield {
-                        peer: BitfieldPeer::Local,
-                        hash,
-                        id: subscription_id,
-                    });
-                    self.bitfields.insert(
-                        (BitfieldPeer::Local, hash),
-                        PeerBlobState::new(subscription_id),
-                    );
-                }
-                self.observers.insert(id, ObserveInfo { hash, ranges });
-            }
-            Command::StopObserveLocal { id } => {
-                let request = self
-                    .observers
-                    .remove(&id)
-                    .context(format!("stop observing unknown local bitfield {id:?}"))?;
-                let removed_hash = request.hash;
-                // unsubscribe from bitfields that have no more subscriptions
-                self.bitfields.retain(|(_peer, hash), state| {
-                    if *hash == removed_hash {
-                        state.subscription_count -= 1;
-                        if state.subscription_count == 0 {
-                            evs.push(Event::UnsubscribeBitfield {
-                                id: state.subscription_id,
-                            });
-                            return false;
-                        }
-                    }
-                    true
-                });
-            }
-        }
-        Ok(())
-    }
-
-    /// Check for completion of a download or of an individual peer download
-    ///
-    /// This must be called after each change of the local bitfield for a hash
-    ///
-    /// In addition to checking for completion, this also create new peer downloads if a peer download is complete and there is more data available for that peer.
-    fn check_completion(
-        &mut self,
-        hash: Hash,
-        just_id: Option<DownloadId>,
-        evs: &mut Vec<Event>,
-    ) -> anyhow::Result<()> {
-        let Some(self_state) = self.bitfields.get_local(hash) else {
-            // we don't have the self state yet, so we can't really decide if we need to download anything at all
-            return Ok(());
-        };
-        let mut completed = vec![];
-        for (id, download) in self.downloads.iter_mut_for_hash(hash) {
-            if just_id.is_some() && just_id != Some(*id) {
-                continue;
-            }
-            // check if the entire download is complete. If this is the case, peer downloads will be cleaned up later
-            if self_state.ranges.is_superset(&download.request.ranges) {
-                // notify the user that the download is complete
-                evs.push(Event::DownloadComplete { id: *id });
-                // remember id for later cleanup
-                completed.push(*id);
-                // no need to look at individual peer downloads in this case
-                continue;
-            }
-            // check if any peer download is complete, and remove it.
-            let mut available = vec![];
-            download.peer_downloads.retain(|peer, peer_download| {
-                if self_state.ranges.is_superset(&peer_download.ranges) {
-                    // stop this peer download.
-                    //
-                    // Might be a noop if the cause for this local change was the same peer download, but we don't know.
-                    evs.push(Event::StopPeerDownload {
-                        id: peer_download.id,
-                    });
-                    // mark this peer as available
-                    available.push(*peer);
-                    false
-                } else {
-                    true
-                }
-            });
-            // reassign the newly available peers without doing a full rebalance
-            if !available.is_empty() {
-                // check if any of the available peers can provide something of the remaining data
-                let mut remaining = &download.request.ranges - &self_state.ranges;
-                // subtract the ranges that are already being taken care of by remaining peer downloads
-                for peer_download in download.peer_downloads.values() {
-                    remaining.difference_with(&peer_download.ranges);
-                }
-                // see what the new peers can do for us
-                let mut candidates = BTreeMap::new();
-                for peer in available {
-                    let Some(peer_state) = self.bitfields.get(&(BitfieldPeer::Remote(peer), hash))
-                    else {
-                        // weird. we should have a bitfield for this peer since it just completed a download
-                        continue;
-                    };
-                    let intersection = &peer_state.ranges & &remaining;
-                    if !intersection.is_empty() {
-                        candidates.insert(peer, intersection);
-                    }
-                }
-                // deduplicate the ranges
-                self.planner.plan(hash, &mut candidates);
-                // start new downloads
-                for (peer, ranges) in candidates {
-                    let id = self.peer_download_id_gen.next();
-                    evs.push(Event::StartPeerDownload {
-                        id,
-                        peer,
-                        hash,
-                        ranges: ranges.clone(),
-                    });
-                    download
-                        .peer_downloads
-                        .insert(peer, PeerDownloadState { id, ranges });
-                }
-            }
-        }
-        // cleanup completed downloads, has to happen later to avoid double mutable borrow
-        for id in completed {
-            self.stop_download(id, evs)?;
-        }
-        Ok(())
-    }
-
-    /// Look at all downloads for a hash and start peer downloads for those that do not have any yet
-    fn start_downloads(
-        &mut self,
-        hash: Hash,
-        just_id: Option<DownloadId>,
-        evs: &mut Vec<Event>,
-    ) -> anyhow::Result<()> {
-        let Some(self_state) = self.bitfields.get_local(hash) else {
-            // we don't have the self state yet, so we can't really decide if we need to download anything at all
-            return Ok(());
-        };
-        for (id, download) in self
-            .downloads
-            .iter_mut_for_hash(hash)
-            .filter(|(_, download)| download.peer_downloads.is_empty())
-        {
-            if just_id.is_some() && just_id != Some(*id) {
-                continue;
-            }
-            let remaining = &download.request.ranges - &self_state.ranges;
-            let mut candidates = BTreeMap::new();
-            for (peer, bitfield) in self.bitfields.remote_for_hash(hash) {
-                let intersection = &bitfield.ranges & &remaining;
-                if !intersection.is_empty() {
-                    candidates.insert(*peer, intersection);
-                }
-            }
-            self.planner.plan(hash, &mut candidates);
-            for (peer, ranges) in candidates {
-                info!("  Starting download from {peer} for {hash} {ranges:?}");
-                let id = self.peer_download_id_gen.next();
-                evs.push(Event::StartPeerDownload {
-                    id,
-                    peer,
-                    hash,
-                    ranges: ranges.clone(),
-                });
-                download
-                    .peer_downloads
-                    .insert(peer, PeerDownloadState { id, ranges });
-            }
-        }
-        Ok(())
-    }
-
-    /// rebalance a single download
-    fn rebalance_download(&mut self, id: DownloadId, evs: &mut Vec<Event>) -> anyhow::Result<()> {
-        let download = self
-            .downloads
-            .by_id_mut(id)
-            .context(format!("rebalancing unknown download {id:?}"))?;
-        download.needs_rebalancing = false;
-        tracing::info!("Rebalancing download {id:?} {:?}", download.request);
-        let hash = download.request.hash;
-        let Some(self_state) = self.bitfields.get_local(hash) else {
-            // we don't have the self state yet, so we can't really decide if we need to download anything at all
-            return Ok(());
-        };
-        let remaining = &download.request.ranges - &self_state.ranges;
-        let mut candidates = BTreeMap::new();
-        for (peer, bitfield) in self.bitfields.remote_for_hash(hash) {
-            let intersection = &bitfield.ranges & &remaining;
-            if !intersection.is_empty() {
-                candidates.insert(*peer, intersection);
-            }
-        }
-        self.planner.plan(hash, &mut candidates);
-        info!(
-            "Stopping {} old peer downloads",
-            download.peer_downloads.len()
-        );
-        for (_, state) in &download.peer_downloads {
-            // stop all downloads
-            evs.push(Event::StopPeerDownload { id: state.id });
-        }
-        info!("Creating {} new peer downloads", candidates.len());
-        download.peer_downloads.clear();
-        for (peer, ranges) in candidates {
-            info!("  Starting download from {peer} for {hash} {ranges:?}");
-            let id = self.peer_download_id_gen.next();
-            evs.push(Event::StartPeerDownload {
-                id,
-                peer,
-                hash,
-                ranges: ranges.clone(),
-            });
-            download
-                .peer_downloads
-                .insert(peer, PeerDownloadState { id, ranges });
-        }
-        Ok(())
-    }
 }
 
 fn total_chunks(chunks: &ChunkRanges) -> Option<u64> {
@@ -1210,42 +361,9 @@ impl Downloader {
     }
 }
 
-#[derive(Debug)]
-struct ObserveInfo {
-    hash: Hash,
-    ranges: ChunkRanges,
-}
-
-#[derive(Debug, Default)]
-struct Observers {
-    by_hash_and_id: BTreeMap<Hash, BTreeMap<ObserveId, ObserveInfo>>,
-}
-
-impl Observers {
-    fn insert(&mut self, id: ObserveId, request: ObserveInfo) {
-        self.by_hash_and_id
-            .entry(request.hash)
-            .or_default()
-            .insert(id, request);
-    }
-
-    fn remove(&mut self, id: &ObserveId) -> Option<ObserveInfo> {
-        for requests in self.by_hash_and_id.values_mut() {
-            if let Some(request) = requests.remove(id) {
-                return Some(request);
-            }
-        }
-        None
-    }
-
-    fn get_by_hash(&self, hash: &Hash) -> Option<&BTreeMap<ObserveId, ObserveInfo>> {
-        self.by_hash_and_id.get(hash)
-    }
-}
-
 /// An user-facing command
 #[derive(Debug)]
-enum UserCommand {
+pub(super) enum UserCommand {
     Download {
         request: DownloadRequest,
         done: tokio::sync::oneshot::Sender<()>,
@@ -1255,221 +373,6 @@ enum UserCommand {
         send: tokio::sync::mpsc::Sender<ObserveEvent>,
     },
 }
-
-struct DownloaderActor<S> {
-    local_pool: LocalPool,
-    endpoint: Endpoint,
-    command_rx: mpsc::Receiver<Command>,
-    command_tx: mpsc::Sender<Command>,
-    state: DownloaderState,
-    store: S,
-    discovery: BoxedContentDiscovery,
-    subscribe_bitfield: BoxedBitfieldSubscription,
-    download_futs: BTreeMap<DownloadId, tokio::sync::oneshot::Sender<()>>,
-    peer_download_tasks: BTreeMap<PeerDownloadId, local_pool::Run<()>>,
-    discovery_tasks: BTreeMap<DiscoveryId, AbortOnDropHandle<()>>,
-    bitfield_subscription_tasks: BTreeMap<BitfieldSubscriptionId, AbortOnDropHandle<()>>,
-    /// Id generator for download ids
-    download_id_gen: IdGenerator<DownloadId>,
-    /// Id generator for observe ids
-    observe_id_gen: IdGenerator<ObserveId>,
-    /// Observers
-    observers: BTreeMap<ObserveId, tokio::sync::mpsc::Sender<ObserveEvent>>,
-    /// The time when the actor was started, serves as the epoch for time messages to the state machine
-    start: Instant,
-}
-
-impl<S: Store> DownloaderActor<S> {
-    fn new(
-        endpoint: Endpoint,
-        store: S,
-        discovery: BoxedContentDiscovery,
-        subscribe_bitfield: BoxedBitfieldSubscription,
-        local_pool: LocalPool,
-        planner: Box<dyn DownloadPlanner>,
-    ) -> Self {
-        let (send, recv) = mpsc::channel(256);
-        Self {
-            local_pool,
-            endpoint,
-            state: DownloaderState::new(planner),
-            store,
-            discovery,
-            subscribe_bitfield,
-            peer_download_tasks: BTreeMap::new(),
-            discovery_tasks: BTreeMap::new(),
-            bitfield_subscription_tasks: BTreeMap::new(),
-            download_futs: BTreeMap::new(),
-            command_tx: send,
-            command_rx: recv,
-            download_id_gen: Default::default(),
-            observe_id_gen: Default::default(),
-            observers: Default::default(),
-            start: Instant::now(),
-        }
-    }
-
-    async fn run(mut self, mut channel: mpsc::Receiver<UserCommand>) {
-        let mut ticks = tokio::time::interval(Duration::from_millis(100));
-        loop {
-            trace!("downloader actor tick");
-            tokio::select! {
-                biased;
-                Some(cmd) = channel.recv() => {
-                    debug!("user command {cmd:?}");
-                    match cmd {
-                        UserCommand::Download {
-                            request, done,
-                        } => {
-                            let id = self.download_id_gen.next();
-                            self.download_futs.insert(id, done);
-                            self.command_tx.send(Command::StartDownload { request, id }).await.ok();
-                        }
-                        UserCommand::Observe { request, send } => {
-                            let id = self.observe_id_gen.next();
-                            self.command_tx.send(Command::ObserveLocal { id, hash: request.hash, ranges: request.ranges }).await.ok();
-                            self.observers.insert(id, send);
-                        }
-                    }
-                },
-                Some(cmd) = self.command_rx.recv() => {
-                    let evs = self.state.apply(cmd);
-                    for ev in evs {
-                        self.handle_event(ev);
-                    }
-                },
-                _ = ticks.tick() => {
-                    let time = self.start.elapsed();
-                    self.command_tx.send(Command::Tick { time }).await.ok();
-                    // clean up dropped futures
-                    //
-                    // todo: is there a better mechanism than periodic checks?
-                    // I don't want some cancellation token rube goldberg machine.
-                    let mut to_delete = vec![];
-                    for (id, fut) in self.download_futs.iter() {
-                        if fut.is_closed() {
-                            to_delete.push(*id);
-                            self.command_tx.send(Command::StopDownload { id: *id }).await.ok();
-                        }
-                    }
-                    for id in to_delete {
-                        self.download_futs.remove(&id);
-                    }
-                    // clean up dropped observers
-                    let mut to_delete = vec![];
-                    for (id, sender) in self.observers.iter() {
-                        if sender.is_closed() {
-                            to_delete.push(*id);
-                            self.command_tx.send(Command::StopObserveLocal { id: *id }).await.ok();
-                        }
-                    }
-                    for id in to_delete {
-                        self.observers.remove(&id);
-                    }
-                },
-            }
-        }
-    }
-
-    fn handle_event(&mut self, ev: Event) {
-        trace!("handle_event {ev:?}");
-        match ev {
-            Event::SubscribeBitfield { peer, hash, id } => {
-                let send = self.command_tx.clone();
-                let mut stream = self.subscribe_bitfield.subscribe(peer, hash);
-                let task = spawn(async move {
-                    while let Some(ev) = stream.next().await {
-                        let cmd = match ev {
-                            BitfieldSubscriptionEvent::Bitfield { ranges } => {
-                                Command::Bitfield { peer, hash, ranges }
-                            }
-                            BitfieldSubscriptionEvent::BitfieldUpdate { added, removed } => {
-                                Command::BitfieldUpdate {
-                                    peer,
-                                    hash,
-                                    added,
-                                    removed,
-                                }
-                            }
-                        };
-                        send.send(cmd).await.ok();
-                    }
-                });
-                self.bitfield_subscription_tasks.insert(id, task);
-            }
-            Event::StartDiscovery { hash, id } => {
-                let send = self.command_tx.clone();
-                let mut stream = self.discovery.find_peers(hash, Default::default());
-                let task = spawn(async move {
-                    // process the infinite discovery stream and send commands
-                    while let Some(peer) = stream.next().await {
-                        println!("peer discovered for hash {hash}: {peer}");
-                        let res = send.send(Command::PeerDiscovered { peer, hash }).await;
-                        if res.is_err() {
-                            // only reason for this is actor task dropped
-                            break;
-                        }
-                    }
-                });
-                self.discovery_tasks.insert(id, task);
-            }
-            Event::StartPeerDownload {
-                id,
-                peer,
-                hash,
-                ranges,
-            } => {
-                let send = self.command_tx.clone();
-                let endpoint = self.endpoint.clone();
-                let store = self.store.clone();
-                let start = self.start;
-                let task = self.local_pool.spawn(move || {
-                    peer_download_task(id, endpoint, store, hash, peer, ranges, send, start)
-                });
-                self.peer_download_tasks.insert(id, task);
-            }
-            Event::UnsubscribeBitfield { id } => {
-                self.bitfield_subscription_tasks.remove(&id);
-            }
-            Event::StopDiscovery { id } => {
-                self.discovery_tasks.remove(&id);
-            }
-            Event::StopPeerDownload { id } => {
-                self.peer_download_tasks.remove(&id);
-            }
-            Event::DownloadComplete { id } => {
-                if let Some(done) = self.download_futs.remove(&id) {
-                    done.send(()).ok();
-                }
-            }
-            Event::LocalBitfield { id, ranges } => {
-                let Some(sender) = self.observers.get(&id) else {
-                    return;
-                };
-                if sender.try_send(ObserveEvent::Bitfield { ranges }).is_err() {
-                    // the observer has been dropped
-                    self.observers.remove(&id);
-                }
-            }
-            Event::LocalBitfieldUpdate { id, added, removed } => {
-                let Some(sender) = self.observers.get(&id) else {
-                    return;
-                };
-                if sender
-                    .try_send(ObserveEvent::BitfieldUpdate { added, removed })
-                    .is_err()
-                {
-                    // the observer has been dropped
-                    self.observers.remove(&id);
-                }
-            }
-            Event::Error { message } => {
-                error!("Error during processing event {}", message);
-            }
-        }
-    }
-}
-
 /// A simple static content discovery mechanism
 #[derive(Debug)]
 pub struct StaticContentDiscovery {
@@ -1495,114 +398,6 @@ impl ContentDiscovery for StaticContentDiscovery {
         let peers = self.info.get(&hash).unwrap_or(&self.default).clone();
         Box::pin(futures_lite::stream::iter(peers).chain(futures_lite::stream::pending()))
     }
-}
-
-async fn peer_download_task<S: Store>(
-    id: PeerDownloadId,
-    endpoint: Endpoint,
-    store: S,
-    hash: Hash,
-    peer: NodeId,
-    ranges: ChunkRanges,
-    sender: mpsc::Sender<Command>,
-    start: Instant,
-) {
-    let result = peer_download(endpoint, store, hash, peer, ranges, &sender, start).await;
-    sender
-        .send(Command::PeerDownloadComplete { id, result })
-        .await
-        .ok();
-}
-
-async fn peer_download<S: Store>(
-    endpoint: Endpoint,
-    store: S,
-    hash: Hash,
-    peer: NodeId,
-    ranges: ChunkRanges,
-    sender: &mpsc::Sender<Command>,
-    start: Instant,
-) -> anyhow::Result<Stats> {
-    info!("Connecting to peer {peer}");
-    let conn = endpoint.connect(peer, crate::ALPN).await?;
-    info!("Got connection to peer {peer}");
-    let spec = RangeSpec::new(ranges);
-    let ranges = RangeSpecSeq::new([spec, RangeSpec::EMPTY]);
-    info!("starting download from {peer} for {hash} {ranges:?}");
-    let request = GetRequest::new(hash, ranges);
-    let initial = crate::get::fsm::start(conn, request);
-    // connect
-    let connected = initial.next().await?;
-    // read the first bytes
-    let ConnectedNext::StartRoot(start_root) = connected.next().await? else {
-        return Err(io::Error::new(io::ErrorKind::Other, "expected start root").into());
-    };
-    let header = start_root.next();
-
-    // get the size of the content
-    let (mut content, size) = header.next().await?;
-    let entry = store.get_or_create(hash, size).await?;
-    let mut writer = entry.batch_writer().await?;
-    let mut batch = Vec::new();
-    // manually loop over the content and yield all data
-    let done = loop {
-        match content.next().await {
-            BlobContentNext::More((next, data)) => {
-                match data? {
-                    BaoContentItem::Parent(parent) => {
-                        batch.push(parent.into());
-                    }
-                    BaoContentItem::Leaf(leaf) => {
-                        let start_chunk = leaf.offset / 1024;
-                        let added =
-                            ChunkRanges::from(ChunkNum(start_chunk)..ChunkNum(start_chunk + 16));
-                        sender
-                            .send(Command::ChunksDownloaded {
-                                time: start.elapsed(),
-                                peer,
-                                hash,
-                                added: added.clone(),
-                            })
-                            .await
-                            .ok();
-                        batch.push(leaf.into());
-                        writer.write_batch(size, std::mem::take(&mut batch)).await?;
-                        sender
-                            .send(Command::BitfieldUpdate {
-                                peer: BitfieldPeer::Local,
-                                hash,
-                                added,
-                                removed: ChunkRanges::empty(),
-                            })
-                            .await
-                            .ok();
-                    }
-                }
-                content = next;
-            }
-            BlobContentNext::Done(done) => {
-                // we are done with the root blob
-                break done;
-            }
-        }
-    };
-    // close the connection even if there is more data
-    let closing = match done.next() {
-        EndBlobNext::Closing(closing) => closing,
-        EndBlobNext::MoreChildren(more) => more.finish(),
-    };
-    // close the connection
-    let stats = closing.next().await?;
-    Ok(stats)
-}
-
-fn spawn<F, T>(f: F) -> AbortOnDropHandle<T>
-where
-    F: Future<Output = T> + Send + 'static,
-    T: Send + 'static,
-{
-    let task = tokio::spawn(f);
-    AbortOnDropHandle::new(task)
 }
 
 /// A bitfield subscription that just returns nothing for local and everything(*) for remote
@@ -1848,7 +643,7 @@ mod tests {
     }
 
     /// Create chunk ranges from an array of u64 ranges
-    fn chunk_ranges(ranges: impl IntoIterator<Item = Range<u64>>) -> ChunkRanges {
+    pub fn chunk_ranges(ranges: impl IntoIterator<Item = Range<u64>>) -> ChunkRanges {
         let mut res = ChunkRanges::empty();
         for range in ranges.into_iter() {
             res |= ChunkRanges::from(ChunkNum(range.start)..ChunkNum(range.end));
@@ -1856,483 +651,21 @@ mod tests {
         res
     }
 
-    fn noop_planner() -> BoxedDownloadPlanner {
+    pub fn noop_planner() -> BoxedDownloadPlanner {
         Box::new(NoopPlanner)
     }
 
     /// Checks if an exact event is present exactly once in a list of events
-    fn has_one_event(evs: &[Event], ev: &Event) -> bool {
+    pub fn has_one_event(evs: &[Event], ev: &Event) -> bool {
         evs.iter().filter(|e| *e == ev).count() == 1
     }
 
-    fn has_all_events(evs: &[Event], evs2: &[&Event]) -> bool {
+    pub fn has_all_events(evs: &[Event], evs2: &[&Event]) -> bool {
         evs2.iter().all(|ev| has_one_event(evs, ev))
     }
 
-    fn has_one_event_matching(evs: &[Event], f: impl Fn(&Event) -> bool) -> bool {
+    pub fn has_one_event_matching(evs: &[Event], f: impl Fn(&Event) -> bool) -> bool {
         evs.iter().filter(|e| f(e)).count() == 1
-    }
-
-    /// Test various things that should produce errors
-    #[test]
-    fn downloader_state_errors() -> TestResult<()> {
-        use BitfieldPeer::*;
-        let _ = tracing_subscriber::fmt::try_init();
-        let peer_a = "1000000000000000000000000000000000000000000000000000000000000000".parse()?;
-        let hash = "0000000000000000000000000000000000000000000000000000000000000001".parse()?;
-        let unknown_hash =
-            "0000000000000000000000000000000000000000000000000000000000000002".parse()?;
-        let mut state = DownloaderState::new(noop_planner());
-        let evs = state.apply(Command::Bitfield {
-            peer: Local,
-            hash,
-            ranges: ChunkRanges::all(),
-        });
-        assert!(
-            has_one_event_matching(&evs, |e| matches!(e, Event::Error { .. })),
-            "adding an open bitfield should produce an error!"
-        );
-        let evs = state.apply(Command::Bitfield {
-            peer: Local,
-            hash: unknown_hash,
-            ranges: ChunkRanges::all(),
-        });
-        assert!(
-            has_one_event_matching(&evs, |e| matches!(e, Event::Error { .. })),
-            "adding an open bitfield for an unknown hash should produce an error!"
-        );
-        let evs = state.apply(Command::ChunksDownloaded {
-            time: Duration::ZERO,
-            peer: peer_a,
-            hash,
-            added: chunk_ranges([0..16]),
-        });
-        assert!(
-            has_one_event_matching(&evs, |e| matches!(e, Event::Error { .. })),
-            "download from unknown peer should lead to an error!"
-        );
-        Ok(())
-    }
-
-    /// Test a simple scenario where a download is started and completed
-    #[test]
-    fn downloader_state_smoke() -> TestResult<()> {
-        use BitfieldPeer::*;
-        let _ = tracing_subscriber::fmt::try_init();
-        let peer_a = "1000000000000000000000000000000000000000000000000000000000000000".parse()?;
-        let hash = "0000000000000000000000000000000000000000000000000000000000000001".parse()?;
-        let mut state = DownloaderState::new(noop_planner());
-        let evs = state.apply(Command::StartDownload {
-            request: DownloadRequest {
-                hash,
-                ranges: chunk_ranges([0..64]),
-            },
-            id: DownloadId(0),
-        });
-        assert!(
-            has_one_event(
-                &evs,
-                &Event::StartDiscovery {
-                    hash,
-                    id: DiscoveryId(0)
-                }
-            ),
-            "starting a download should start a discovery task"
-        );
-        assert!(
-            has_one_event(
-                &evs,
-                &Event::SubscribeBitfield {
-                    peer: Local,
-                    hash,
-                    id: BitfieldSubscriptionId(0)
-                }
-            ),
-            "starting a download should subscribe to the local bitfield"
-        );
-        let initial_bitfield = ChunkRanges::from(ChunkNum(0)..ChunkNum(16));
-        let evs = state.apply(Command::Bitfield {
-            peer: Local,
-            hash,
-            ranges: initial_bitfield.clone(),
-        });
-        assert!(evs.is_empty());
-        assert_eq!(
-            state
-                .bitfields
-                .get_local(hash)
-                .context("bitfield should be present")?
-                .ranges,
-            initial_bitfield,
-            "bitfield should be set to the initial bitfield"
-        );
-        assert_eq!(
-            state
-                .bitfields
-                .get_local(hash)
-                .context("bitfield should be present")?
-                .subscription_count,
-            1,
-            "we have one download interested in the bitfield"
-        );
-        let evs = state.apply(Command::BitfieldUpdate {
-            peer: Local,
-            hash,
-            added: chunk_ranges([16..32]),
-            removed: ChunkRanges::empty(),
-        });
-        assert!(evs.is_empty());
-        assert_eq!(
-            state
-                .bitfields
-                .get_local(hash)
-                .context("bitfield should be present")?
-                .ranges,
-            ChunkRanges::from(ChunkNum(0)..ChunkNum(32)),
-            "bitfield should be updated"
-        );
-        let evs = state.apply(Command::PeerDiscovered { peer: peer_a, hash });
-        assert!(
-            has_one_event(
-                &evs,
-                &Event::SubscribeBitfield {
-                    peer: Remote(peer_a),
-                    hash,
-                    id: 1.into()
-                }
-            ),
-            "adding a new peer for a hash we are interested in should subscribe to the bitfield"
-        );
-        let evs = state.apply(Command::Bitfield {
-            peer: Remote(peer_a),
-            hash,
-            ranges: chunk_ranges([0..64]),
-        });
-        assert!(
-            has_one_event(
-                &evs,
-                &Event::StartPeerDownload {
-                    id: PeerDownloadId(0),
-                    peer: peer_a,
-                    hash,
-                    ranges: chunk_ranges([32..64])
-                }
-            ),
-            "bitfield from a peer should start a download"
-        );
-        // ChunksDownloaded just updates the peer stats
-        let evs = state.apply(Command::ChunksDownloaded {
-            time: Duration::ZERO,
-            peer: peer_a,
-            hash,
-            added: chunk_ranges([32..48]),
-        });
-        assert!(evs.is_empty());
-        // Bitfield update does not yet complete the download
-        let evs = state.apply(Command::BitfieldUpdate {
-            peer: Local,
-            hash,
-            added: chunk_ranges([32..48]),
-            removed: ChunkRanges::empty(),
-        });
-        assert!(evs.is_empty());
-        // ChunksDownloaded just updates the peer stats
-        let evs = state.apply(Command::ChunksDownloaded {
-            time: Duration::ZERO,
-            peer: peer_a,
-            hash,
-            added: chunk_ranges([48..64]),
-        });
-        assert!(evs.is_empty());
-        // Final bitfield update for the local bitfield should complete the download
-        let evs = state.apply(Command::BitfieldUpdate {
-            peer: Local,
-            hash,
-            added: chunk_ranges([48..64]),
-            removed: ChunkRanges::empty(),
-        });
-        assert!(
-            has_one_event_matching(&evs, |e| matches!(e, Event::DownloadComplete { .. })),
-            "download should be completed by the data"
-        );
-        // quick check that everything got cleaned up
-        assert!(state.downloads.by_id.is_empty());
-        assert!(state.bitfields.by_peer_and_hash.is_empty());
-        assert!(state.discovery.is_empty());
-        Ok(())
-    }
-
-    /// Test a scenario where more data becomes available at the remote peer as the download progresses
-    #[test]
-    fn downloader_state_incremental() -> TestResult<()> {
-        use BitfieldPeer::*;
-        let _ = tracing_subscriber::fmt::try_init();
-        let peer_a = "1000000000000000000000000000000000000000000000000000000000000000".parse()?;
-        let hash = "0000000000000000000000000000000000000000000000000000000000000001".parse()?;
-        let mut state = DownloaderState::new(noop_planner());
-        // Start a download
-        state.apply(Command::StartDownload {
-            request: DownloadRequest {
-                hash,
-                ranges: chunk_ranges([0..64]),
-            },
-            id: DownloadId(0),
-        });
-        // Initially, we have nothing
-        state.apply(Command::Bitfield {
-            peer: Local,
-            hash,
-            ranges: ChunkRanges::empty(),
-        });
-        // We have a peer for the hash
-        state.apply(Command::PeerDiscovered { peer: peer_a, hash });
-        // We have a bitfield from the peer
-        let evs = state.apply(Command::Bitfield {
-            peer: Remote(peer_a),
-            hash,
-            ranges: chunk_ranges([0..32]),
-        });
-        assert!(
-            has_one_event(
-                &evs,
-                &Event::StartPeerDownload {
-                    id: 0.into(),
-                    peer: peer_a,
-                    hash,
-                    ranges: chunk_ranges([0..32])
-                }
-            ),
-            "bitfield from a peer should start a download"
-        );
-        // ChunksDownloaded just updates the peer stats
-        state.apply(Command::ChunksDownloaded {
-            time: Duration::ZERO,
-            peer: peer_a,
-            hash,
-            added: chunk_ranges([0..16]),
-        });
-        // Bitfield update does not yet complete the download
-        state.apply(Command::BitfieldUpdate {
-            peer: Local,
-            hash,
-            added: chunk_ranges([0..16]),
-            removed: ChunkRanges::empty(),
-        });
-        // The peer now has more data
-        state.apply(Command::Bitfield {
-            peer: Remote(peer_a),
-            hash,
-            ranges: chunk_ranges([32..64]),
-        });
-        // ChunksDownloaded just updates the peer stats
-        state.apply(Command::ChunksDownloaded {
-            time: Duration::ZERO,
-            peer: peer_a,
-            hash,
-            added: chunk_ranges([16..32]),
-        });
-        // Complete the first part of the download
-        let evs = state.apply(Command::BitfieldUpdate {
-            peer: Local,
-            hash,
-            added: chunk_ranges([16..32]),
-            removed: ChunkRanges::empty(),
-        });
-        // This triggers cancellation of the first peer download and starting a new one for the remaining data
-        assert!(
-            has_one_event(&evs, &Event::StopPeerDownload { id: 0.into() }),
-            "first peer download should be stopped"
-        );
-        assert!(
-            has_one_event(
-                &evs,
-                &Event::StartPeerDownload {
-                    id: 1.into(),
-                    peer: peer_a,
-                    hash,
-                    ranges: chunk_ranges([32..64])
-                }
-            ),
-            "second peer download should be started"
-        );
-        // ChunksDownloaded just updates the peer stats
-        state.apply(Command::ChunksDownloaded {
-            time: Duration::ZERO,
-            peer: peer_a,
-            hash,
-            added: chunk_ranges([32..64]),
-        });
-        // Final bitfield update for the local bitfield should complete the download
-        let evs = state.apply(Command::BitfieldUpdate {
-            peer: Local,
-            hash,
-            added: chunk_ranges([32..64]),
-            removed: ChunkRanges::empty(),
-        });
-        assert!(
-            has_all_events(
-                &evs,
-                &[
-                    &Event::StopPeerDownload { id: 1.into() },
-                    &Event::DownloadComplete { id: 0.into() },
-                    &Event::UnsubscribeBitfield { id: 0.into() },
-                    &Event::StopDiscovery { id: 0.into() },
-                ]
-            ),
-            "download should be completed by the data"
-        );
-        println!("{evs:?}");
-        Ok(())
-    }
-
-    #[test]
-    fn downloader_state_multiple_downloads() -> testresult::TestResult<()> {
-        use BitfieldPeer::*;
-        // Use a constant hash (the same style as used in other tests).
-        let hash = "0000000000000000000000000000000000000000000000000000000000000001".parse()?;
-        // Create a downloader state with a no‐op planner.
-        let mut state = DownloaderState::new(noop_planner());
-
-        // --- Start the first (ongoing) download.
-        // Request a range from 0..64.
-        let download0 = DownloadId(0);
-        let req0 = DownloadRequest {
-            hash,
-            ranges: chunk_ranges([0..64]),
-        };
-        let evs0 = state.apply(Command::StartDownload {
-            request: req0,
-            id: download0,
-        });
-        // When starting the download, we expect a discovery task to be started
-        // and a subscription to the local bitfield to be requested.
-        assert!(
-            has_one_event(
-                &evs0,
-                &Event::StartDiscovery {
-                    hash,
-                    id: DiscoveryId(0)
-                }
-            ),
-            "download0 should start discovery"
-        );
-        assert!(
-            has_one_event(
-                &evs0,
-                &Event::SubscribeBitfield {
-                    peer: Local,
-                    hash,
-                    id: BitfieldSubscriptionId(0)
-                }
-            ),
-            "download0 should subscribe to the local bitfield"
-        );
-
-        // --- Simulate some progress for the first download.
-        // Let’s say only chunks 0..32 are available locally.
-        let evs1 = state.apply(Command::Bitfield {
-            peer: Local,
-            hash,
-            ranges: chunk_ranges([0..32]),
-        });
-        // No completion event should be generated for download0 because its full range 0..64 is not yet met.
-        assert!(
-            evs1.is_empty(),
-            "Partial bitfield update should not complete download0"
-        );
-
-        // --- Start a second download for the same hash.
-        // This new download only requires chunks 0..32 which are already available.
-        let download1 = DownloadId(1);
-        let req1 = DownloadRequest {
-            hash,
-            ranges: chunk_ranges([0..32]),
-        };
-        let evs2 = state.apply(Command::StartDownload {
-            request: req1,
-            id: download1,
-        });
-        // Because the local bitfield (0..32) is already a superset of the new download’s request,
-        // a DownloadComplete event for download1 should be generated immediately.
-        assert!(
-            has_one_event(&evs2, &Event::DownloadComplete { id: download1 }),
-            "New download should complete immediately"
-        );
-
-        // --- Verify state:
-        // The ongoing download (download0) should still be present in the state,
-        // while the newly completed download (download1) is removed.
-        assert!(
-            state.downloads.contains_key(&download0),
-            "download0 should still be active"
-        );
-        assert!(
-            !state.downloads.contains_key(&download1),
-            "download1 should have been cleaned up after completion"
-        );
-
-        Ok(())
-    }
-
-    /// Test a scenario where more data becomes available at the remote peer as the download progresses
-    #[test]
-    fn downloader_state_drop() -> TestResult<()> {
-        use BitfieldPeer::*;
-        let _ = tracing_subscriber::fmt::try_init();
-        let peer_a = "1000000000000000000000000000000000000000000000000000000000000000".parse()?;
-        let hash = "0000000000000000000000000000000000000000000000000000000000000001".parse()?;
-        let mut state = DownloaderState::new(noop_planner());
-        // Start a download
-        state.apply(Command::StartDownload {
-            request: DownloadRequest {
-                hash,
-                ranges: chunk_ranges([0..64]),
-            },
-            id: 0.into(),
-        });
-        // Initially, we have nothing
-        state.apply(Command::Bitfield {
-            peer: Local,
-            hash,
-            ranges: ChunkRanges::empty(),
-        });
-        // We have a peer for the hash
-        state.apply(Command::PeerDiscovered { peer: peer_a, hash });
-        // We have a bitfield from the peer
-        let evs = state.apply(Command::Bitfield {
-            peer: Remote(peer_a),
-            hash,
-            ranges: chunk_ranges([0..32]),
-        });
-        assert!(
-            has_one_event(
-                &evs,
-                &Event::StartPeerDownload {
-                    id: 0.into(),
-                    peer: peer_a,
-                    hash,
-                    ranges: chunk_ranges([0..32])
-                }
-            ),
-            "bitfield from a peer should start a download"
-        );
-        // Sending StopDownload should stop the download and all associated tasks
-        // This is what happens (delayed) when the user drops the download future
-        let evs = state.apply(Command::StopDownload { id: 0.into() });
-        assert!(has_one_event(
-            &evs,
-            &Event::StopPeerDownload { id: 0.into() }
-        ));
-        assert!(has_one_event(
-            &evs,
-            &Event::UnsubscribeBitfield { id: 0.into() }
-        ));
-        assert!(has_one_event(
-            &evs,
-            &Event::UnsubscribeBitfield { id: 1.into() }
-        ));
-        assert!(has_one_event(&evs, &Event::StopDiscovery { id: 0.into() }));
-        Ok(())
     }
 
     #[tokio::test]
@@ -2355,7 +688,7 @@ mod tests {
             .discovery(discovery)
             .bitfield_subscription(bitfield_subscription)
             .build();
-        tokio::time::sleep(Duration::from_secs(2)).await;
+        tokio::time::sleep(Duration::from_secs(1)).await;
         let fut = downloader.download(DownloadRequest {
             hash,
             ranges: chunk_ranges([0..1]),
